pydantic model fix

+++ app/router.py
@@ -45,7 +45,7 @@ async def payment_notification(
     session: DataInteraction = response.app.state.session
     scheduler: AsyncIOScheduler = response.app.state.scheduler
     order_storage: TTLCache = response.app.state.order_storage
-    if not check_signature_result(int(InvId), int(OutSum), SignatureValue, user_id=Shp_userId, order_id=Shp_orderId):
+    if not check_signature_result(int(InvId), float(OutSum), SignatureValue, user_id=Shp_userId, order_id=Shp_orderId):
         return "bad sign"
     answer = f'OK{InvId}'
     user_id = int(Shp_userId)
